test: collect more diagnostics on failure (#222)

+++ testing/src/main/kotlin/org/neo4j/connectors/kafka/testing/source/Neo4jSourceExtension.kt
@@ -32,6 +32,7 @@ import org.junit.jupiter.api.extension.ExtensionConfigurationException
 import org.junit.jupiter.api.extension.ExtensionContext
 import org.junit.jupiter.api.extension.ParameterContext
 import org.junit.jupiter.api.extension.ParameterResolver
+import org.junit.jupiter.api.extension.TestExecutionExceptionHandler
 import org.neo4j.connectors.kafka.testing.AnnotationSupport
 import org.neo4j.connectors.kafka.testing.AnnotationValueResolver
 import org.neo4j.connectors.kafka.testing.DatabaseSupport.createDatabase
@@ -59,7 +60,12 @@ internal class Neo4jSourceExtension(
         { properties, topic ->
           ConsumerResolver.getSubscribedConsumer(properties, topic)
         },
-) : ExecutionCondition, BeforeEachCallback, AfterEachCallback, ParameterResolver {
+) :
+    ExecutionCondition,
+    BeforeEachCallback,
+    AfterEachCallback,
+    ParameterResolver,
+    TestExecutionExceptionHandler {
 
   private val log: Logger = LoggerFactory.getLogger(Neo4jSourceExtension::class.java)
 
@@ -80,6 +86,8 @@ internal class Neo4jSourceExtension(
   private lateinit var session: Session
 
   private lateinit var neo4jDatabase: String
+
+  private var testFailed: Boolean = false
 
   private val brokerExternalHost =
       AnnotationValueResolver(Neo4jSource::brokerExternalHost, envAccessor)
@@ -173,18 +181,32 @@ internal class Neo4jSourceExtension(
             payloadMode = keyValueConverterResolver.resolvePayloadMode(context))
 
     source.register(kafkaConnectExternalUri.resolve(sourceAnnotation))
+    log.info("registered source connector with name {}", source.name)
     topicRegistry.log()
+
+    testFailed = false
+  }
+
+  override fun handleTestExecutionException(context: ExtensionContext?, throwable: Throwable) {
+    // we do not drop database if the test fails
+    testFailed = true
+
+    throw throwable
   }
 
   override fun afterEach(context: ExtensionContext?) {
     source.unregister()
     if (this::driver.isInitialized) {
-      driver.session(SessionConfig.forDatabase("system")).use { it.dropDatabase(neo4jDatabase) }
+      if (!testFailed) {
+        driver.session(SessionConfig.forDatabase("system")).use { it.dropDatabase(neo4jDatabase) }
+      }
       session.close()
       driver.close()
     } else {
-      createDriver().use { dr ->
-        dr.session(SessionConfig.forDatabase("system")).use { it.dropDatabase(neo4jDatabase) }
+      if (!testFailed) {
+        createDriver().use { dr ->
+          dr.session(SessionConfig.forDatabase("system")).use { it.dropDatabase(neo4jDatabase) }
+        }
       }
     }
   }
@@ -227,6 +249,7 @@ internal class Neo4jSourceExtension(
 
       // want to make sure that CDC is functional before running the test
       if (sourceAnnotation.strategy == SourceStrategy.CDC) {
+        log.info("waiting cdc to be available")
         runBlocking {
           eventually(30.seconds) {
             driver.session(SessionConfig.forDatabase(neo4jDatabase)).use { session ->
@@ -240,11 +263,12 @@ internal class Neo4jSourceExtension(
 
                 count shouldBe 0
               } catch (e: Exception) {
-                log.trace("error received while waiting for cdc to be available", e)
+                log.debug("error received while waiting for cdc to be available", e)
               }
             }
           }
         }
+        log.info("cdc is available")
       }
     }
   }
